NXP backend: Re-organize Neutron backend tests.
